refactor: Edit package structure
