+++ src/main/java/com/example/server/config/oauth/PrincipalOauth2UserService.java
@@ -14,8 +14,8 @@ import com.example.server.config.oauth.provider.OAuth2Provider;
 import com.example.server.config.oauth.provider.OAuth2UserInfo;
 import com.example.server.domain.member.Member;
 import com.example.server.domain.member.RoleType;
-import com.example.server.repository.member.MemberQueryRepository;
-import com.example.server.repository.member.MemberRepository;
+import com.example.server.persistence.member.MemberQueryRepository;
+import com.example.server.persistence.member.MemberRepository;
 
 import lombok.AllArgsConstructor;
