--- src/main/java/com/example/server/exception/CustomException.java
@@ -1,10 +1,0 @@
-package com.example.server.exception;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class CustomException extends RuntimeException {
-	private final ErrorCode errorCode;
-}
