+++ src/main/java/com/example/server/persistence/image/ImageRepository.java
@@ -1,14 +1,14 @@
-package com.example.server.repository.member;
+package com.example.server.persistence.image;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.example.server.domain.member.Member;
+import com.example.server.domain.image.Image;
 
 @Repository
 @Transactional
-public interface MemberRepository extends JpaRepository<Member, Long> {
+public interface ImageRepository extends JpaRepository<Image, Long> {
 }
