+++ src/main/java/com/example/server/persistence/pet/PetQueryRepository.java
@@ -1,14 +1,12 @@
-package com.example.server.repository.image;
+package com.example.server.persistence.pet;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.example.server.domain.image.Image;
+import com.example.server.domain.pet.Pet;
 
 @Repository
-@Transactional
-public interface ImageRepository extends JpaRepository<Image, Long> {
+@Transactional(readOnly = true)
+public interface PetQueryRepository extends JpaRepository<Pet, Long> {
 }
-
-
