+++ src/main/java/com/example/server/util/S3Uploader.java
@@ -11,7 +11,7 @@ import org.springframework.web.multipart.MultipartFile;
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.model.ObjectMetadata;
 import com.example.server.domain.image.Image;
-import com.example.server.repository.image.ImageRepository;
+import com.example.server.persistence.image.ImageRepository;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
